libraries: SW I2C: support for custom clock rates
Original library had hard-coded delay.
Modified it so that it has the same default delay but also adds an
option during the creation of the instance to change the clock rate to a
custom one.

The motivation for this change was that the HiFive1 can run on higher
clocks than the hard-coded delay and originally there was no way to
change the delay without modifying the library.

Signed-off-by: Roi Weisfeld <Roi.Weisfeld@wdc.com>